rename: 파일 및 폴더명 정리 / 컴포넌트 세분화
