+++ src/components/CoinList/CoinListItem.tsx
@@ -3,8 +3,8 @@ import { useSetRecoilState } from "recoil";
 import { coinSelect } from "recoil/atoms";
 import styled from "styled-components";
 import CoinPrice from "./CoinPrice";
-import CoinChange from "./CoinChange";
-import AccTradePrice from "./AccTradePirce";
+import CoinCompare from "./CoinCompare";
+import CoinTradeVolume from "./CoinTradeVolume";
 import CoinName from "./CoinName";
 import { market, ticker } from "types/types";
 
@@ -24,8 +24,8 @@ const CoinListItem = ({ item, coinMarkets }: ListItemProps) => {
     >
       <CoinName koreanName={coinMarkets.korean_name} marketCode={coinMarkets.market} />
       <CoinPrice price={item.trade_price} change={item.change} />
-      <CoinChange change={item.change} rate={item.signed_change_rate} price={item.signed_change_price} />
-      <AccTradePrice price={item.acc_trade_price_24h} />
+      <CoinCompare change={item.change} rate={item.signed_change_rate} price={item.signed_change_price} />
+      <CoinTradeVolume price={item.acc_trade_price_24h} />
     </CoinList>
   );
 };
