+++ src/components/CoinList/CoinPrice.tsx
