+++ src/components/portfolio/PortfolioList.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect } from "react";
 import axios from "axios";
 import { useRecoilState, useRecoilValue } from "recoil";
 import { coinTickers } from "recoil/atoms";
